Fix bug with encoders fetching next work item

+++ src/main/java/net/celestialdata/plexbot/apis/EncodingWorkResource.java
@@ -28,7 +28,11 @@ public class EncodingWorkResource {
     @GET
     @Path("/{id}")
     public EncodingWorkItem get(@PathParam("id") int id) {
-        return encodingWorkItemDao.get(id);
+        if (encodingWorkItemDao.exists(id)) {
+            return encodingWorkItemDao.get(id);
+        } else {
+            throw new NotFoundException();
+        }
     }
 
     @POST
@@ -39,12 +43,20 @@ public class EncodingWorkResource {
     @PUT
     @Path("/{id}")
     public EncodingWorkItem update(@PathParam("id") int id, @QueryParam("progress") String progress) {
-        return encodingWorkItemDao.update(id, progress);
+        if (encodingWorkItemDao.exists(id)) {
+            return encodingWorkItemDao.update(id, progress);
+        } else {
+            throw new NotFoundException();
+        }
     }
 
     @DELETE
     @Path("/{id}")
     public void delete(@PathParam("id") int id) {
-        encodingWorkItemDao.delete(id);
+        if (encodingWorkItemDao.exists(id)) {
+            encodingWorkItemDao.delete(id);
+        } else {
+            throw new NotFoundException();
+        }
     }
 }
